Bug 1598649 - split into 'yarn release' and '.. release:publish'
The former just sets up the changelog and pushes to git.  The latter
publishes the resulting release artifacts (and thus requires
credentials).
