+++ changelog/bug-1598649.md
@@ -1,0 +1,4 @@
+level: silent
+reference: bug 1598649
+---
+Taskcluster releases are now performed in two parts: first in the Git repository, and then publishing all of the various artifacts of the release.
