+++ package.json
@@ -171,6 +171,7 @@
     "test:cleanup": "yarn workspace taskcluster-auth test:cleanup",
     "build": "node infrastructure/tooling/src/main.js build",
     "release": "node infrastructure/tooling/src/main.js release",
+    "release:publish": "node infrastructure/tooling/src/main.js release:publish",
     "generate": "node infrastructure/tooling/src/main.js generate",
     "changelog": "node infrastructure/tooling/src/main.js changelog",
     "changelog:show": "node infrastructure/tooling/src/main.js changelog:show",
